Add `RelationshipAnarchy` flag
The SVG heart path was created by me.

The alternative black corner was not used due to be visually noisy at a small scale, the heart is already pretty detailed.

The thickness and positioning of the heart was changed to be better visible at small scale compared to the reference flag.

resolves: https://github.com/Yellow-Dog-Man/Resonite-Issues/issues/4640
reference flag: https://lgbtqia.wiki/wiki/File:Relationship_anarchy.png